moved all factories to backend/tests and refactored it
