+++ backend/tests/integrated/template_module/services/commands/test_commands.py
@@ -4,8 +4,7 @@ from modules.common.message_bus import MessageBus
 from modules.template_module.domain.commands import SetTemplateValue
 from modules.template_module.services import set_template_value
 
-from ..... import fakers
-from .....unit import factories
+from ..... import factories, fakers
 from ....utils import TestThread
 
 
@@ -14,7 +13,7 @@ def test_concurrent_template_updates_are_handled(
     message_bus: MessageBus,
 ):
     # Given
-    template_entity = factories.TemplateEntityFactory.create()
+    template_entity = factories.TemplateEntityFactory().create()
     final_template_value = fakers.fake_template_value()
     unit_of_work = fake_template_unit_of_work_factory(
         initial_templates=[template_entity]
