+++ backend/tests/unit/template_module/domain/test_entities/test_template.py
@@ -5,13 +5,12 @@ from modules.template_module.domain.entities import set_template_value
 from modules.template_module.domain.exceptions import InvalidTemplateValue
 from modules.template_module.domain.value_objects import TemplateId, TemplateValue
 
-from ..... import fakers
-from .... import factories
+from ..... import factories, fakers
 
 
 def test_set_template_value_sets_value_when_valid_value():
     # Given
-    template = factories.TemplateEntityFactory.create()
+    template = factories.TemplateEntityFactory().create()
     value = fakers.fake_template_value()
 
     # When
@@ -22,7 +21,7 @@ def test_set_template_value_sets_value_when_valid_value():
 
 
 def test_set_template_value_raises_exception_when_invalid_value():
-    template = factories.TemplateEntityFactory.create()
+    template = factories.TemplateEntityFactory().create()
     timestamp_before_setting_value = template.timestamp
 
     with pytest.raises(InvalidTemplateValue):
@@ -33,7 +32,7 @@ def test_set_template_value_raises_exception_when_invalid_value():
 
 def test_set_template_value_method_value_sets_value_when_valid_value():
     # Given
-    template = factories.TemplateEntityFactory.create()
+    template = factories.TemplateEntityFactory().create()
     value = fakers.fake_template_value()
     timestamp_before_setting_value = template.timestamp
 
@@ -46,7 +45,7 @@ def test_set_template_value_method_value_sets_value_when_valid_value():
 
 
 def test_set_template_method_raises_exception_when_invalid_value():
-    template = factories.TemplateEntityFactory.create()
+    template = factories.TemplateEntityFactory().create()
     timestamp_before_setting_value = template.timestamp
 
     with pytest.raises(InvalidTemplateValue):
