+++ backend/tests/unit/template_module/services/queries/test_queries.py
@@ -9,15 +9,14 @@ from modules.template_module.services.queries.mappers import (
     map_template_entity_to_output_dto,
 )
 
-from ..... import fakers
-from .... import factories
+from ..... import factories, fakers
 
 
 def test_get_template_returns_output_dto_when_template_exists(
     fake_template_query_repository_factory: Callable,
 ):
     # Given
-    template = factories.TemplateEntityFactory.create()
+    template = factories.TemplateEntityFactory().create()
     query_repository = fake_template_query_repository_factory(
         initial_templates=[template]
     )
@@ -50,7 +49,7 @@ def test_list_templates_lists_all_templates(
     fake_template_query_repository_factory: Callable,
 ):
     # Given
-    templates = [factories.TemplateEntityFactory.create()]
+    templates = [factories.TemplateEntityFactory().create()]
     query_repository = fake_template_query_repository_factory(
         initial_templates=templates
     )
